@@ -36,23 +36,27 @@ dependencies:
       path: plugins/menubar
 
   cupertino_icons: ^1.0.5
-  supabase_flutter: ^1.2.2
+  supabase_flutter: ^1.3.1
   supabase_auth_ui: ^0.0.1
-  mime: ^1.0.3
+  mime: ^1.0.4
   bitsdojo_window: ^0.1.5
 
   percent_indicator: ^4.2.2
   provider: ^6.0.5
   animations: ^2.0.7
   animated_text_kit: ^4.2.2
-  shared_preferences: ^2.0.15
+  shared_preferences: ^2.0.17
   intl: ^0.18.0
-  flutter_animate: ^2.0.1
+  flutter_animate: ^3.0.0
   sticky_headers: ^0.3.0+2
-  file_picker: ^5.2.4
+  file_picker: ^5.2.5
   enum_to_string: ^2.0.1
   loading_animation_widget: ^1.2.0+4
   flutter_window_close: ^0.2.2
+  uni_links_desktop: ^0.1.4
+  context_menus: ^1.0.2
+  auto_size_text: ^3.0.0
+  parent_child_checkbox: ^0.0.3
 
 dev_dependencies:
   flutter_launcher_icons: ^0.11.0
@@ -93,6 +97,7 @@ flutter:
    - assets/
    - assets/stamp.png
    - assets/logo/logo.png
+   - assets/classes/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
