@@ -1,19 +1,22 @@
 import 'package:enum_to_string/enum_to_string.dart';
+import 'package:maple_daily_tracker/models/maple-class.dart';
 
 import 'action-section.dart';
 import 'action-type.dart';
 
 class Character {
-  final int id;
+  final int? id;
+  final MapleClass? classId;
   final String name;
   final int? order;
-  final DateTime? createdOn;
+  final DateTime createdOn;
   final String subjectId;
   final List<int> hiddenSections;
   Map<ActionType, ActionSection> sections;
 
   Character(
       {required this.id,
+      this.classId,
       required this.name,
       required this.order,
       required this.createdOn,
@@ -23,11 +26,16 @@ class Character {
 
   Character.fromJson(Map<String, dynamic> json)
       : id = json['id'],
+        classId = json['class_id'] == null
+            ? null
+            : MapleClass.values[json['class_id']],
         name = json['name'],
         order = json['order'],
         createdOn = DateTime.parse(json['created']),
         subjectId = json['subject_id'],
-        hiddenSections = json['hidden_sections'] == null ? [] : List<int>.from(json['hidden_sections']),
+        hiddenSections = json['hidden_sections'] == null
+            ? []
+            : List<int>.from(json['hidden_sections']),
         sections = {} {
     sections = {
       ActionType.dailies: ActionSection.empty(ActionType.dailies,
@@ -39,10 +47,20 @@ class Character {
     };
   }
 
+  Map<String, dynamic> toMap() => {
+        'name': name,
+        'order': order,
+        'class_id': classId == null ? null : classId!.index,
+        'created': createdOn.toIso8601String(),
+        'subject_id': subjectId,
+        'hidden_sections': hiddenSections
+      };
+
   void hideSections() {
     var actionTypes = ActionType.values;
     for (var i = 0; i < actionTypes.length; i++) {
-      sections[actionTypes[i]]?.isActive = !hiddenSections.contains(actionTypes[i].index);
+      sections[actionTypes[i]]?.isActive =
+          !hiddenSections.contains(actionTypes[i].index);
     }
   }
 
