@@ -1,9 +1,18 @@
 import 'package:maple_daily_tracker/models/action-type.dart';
+import 'package:maple_daily_tracker/models/maple-class.dart';
 
-extension EnumExtension on ActionType {
-  String toProperName() {
+extension ActionExtension on ActionType {
+  String get name {
     final beforeCapitalLetter = RegExp(r"(?=[A-Z])");
     var parts = this.toString().split('.').last.split(beforeCapitalLetter);
     return parts.join(' ').toLowerCase();
+  }
+}
+
+extension ClassExtension on MapleClass {
+  String get name {
+    final beforeCapitalLetter = RegExp(r"(?=[A-Z])");
+    var parts = this.toString().split('.').last.split(beforeCapitalLetter);
+    return parts.join(' ');
   }
 }
