@@ -3,15 +3,17 @@ import 'package:flutter/material.dart';
 extension ShowSnackBar on BuildContext {
   void showSnackBar({
     required String message,
-    Color backgroundColor = Colors.white,
+    SnackBarAction? action,
+    Color backgroundColor = Colors.blueAccent,
   }) {
     ScaffoldMessenger.of(this).showSnackBar(SnackBar(
       content: Text(message),
       backgroundColor: backgroundColor,
+      action: action,
     ));
   }
 
   void showErrorSnackBar({required String message}) {
-    showSnackBar(message: message, backgroundColor: Colors.red);
+    showSnackBar(message: message, backgroundColor: Colors.redAccent);
   }
 }
