@@ -1,7 +1,7 @@
 import 'package:maple_daily_tracker/extensions/datetime_extensions.dart';
 
 class ResetHelper {
-  Duration calcResetTime({int days = 1}) {
+  static Duration calcResetTime({int days = 1}) {
     var utcNow = DateTime.now().toUtc();
     var resetDate = utcNow.add(Duration(days: days));
 
@@ -9,12 +9,12 @@ class ResetHelper {
     return resetDate.difference(utcNow);
   }
 
-  Duration calcWeeklyResetTime({int resetDay = DateTime.thursday}) {
+  static Duration calcWeeklyResetTime({int resetDay = DateTime.thursday}) {
     int days = daysTillWeekly(resetDay);
     return calcResetTime(days: days);
   }
 
-  int daysTillWeekly(int resetDay) {
+  static int daysTillWeekly(int resetDay) {
     var utcNow = DateTime.now().toUtc();
     var dayOfWeek = utcNow.weekday;
 
