@@ -1,8 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:maple_daily_tracker/providers/tracker.dart';
+import 'package:provider/provider.dart';
 import 'package:supabase_auth_ui/supabase_auth_ui.dart';
 
 class SignInScreen extends StatelessWidget {
-  const SignInScreen({Key? key, required this.onSignUp, required this.onForgotPassword}) : super(key: key);
+  const SignInScreen(
+      {Key? key, required this.onSignUp, required this.onForgotPassword})
+      : super(key: key);
 
   final void Function() onSignUp;
   final void Function() onForgotPassword;
@@ -15,24 +19,22 @@ class SignInScreen extends StatelessWidget {
       children: [
         SupaEmailAuth(
           authAction: SupaAuthAction.signIn,
-          onSuccess: (response) {
-            print(response.user);
+          onSuccess: (authResponse) {
+            context.read<TrackerModel>().fetchProfileInfo(authResponse.user!.id);
           },
         ),
         TextButton(
-          child: const Text(
-            'Forgot Password? Click here',
-            style: TextStyle(fontWeight: FontWeight.bold),
-          ),
-          onPressed: onForgotPassword
-        ),
+            child: const Text(
+              'Forgot Password? Click here',
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
+            onPressed: onForgotPassword),
         TextButton(
-          child: const Text(
-            'Don\'t have an account? Sign Up',
-            style: TextStyle(fontWeight: FontWeight.bold),
-          ),
-          onPressed: onSignUp
-        ),
+            child: const Text(
+              'Don\'t have an account? Sign Up',
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
+            onPressed: onSignUp),
       ],
     );
   }
