@@ -1,9 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:maple_daily_tracker/constants.dart';
+import 'package:maple_daily_tracker/extensions/snackbar_extensions.dart';
 import 'package:supabase_auth_ui/supabase_auth_ui.dart';
 
 class SignUpScreen extends StatelessWidget {
-  const SignUpScreen({Key? key, required this.onSignIn}) : super(key: key);
+  const SignUpScreen({Key? key, required this.onSignIn, required this.onSignUpSuccess}) : super(key: key);
 
+  final void Function() onSignUpSuccess;
   final void Function() onSignIn;
 
   @override
@@ -14,9 +17,12 @@ class SignUpScreen extends StatelessWidget {
       children: [
         SupaEmailAuth(
           authAction: SupaAuthAction.signUp,
+          redirectUrl: SUPABASE_REDIRECT_URL,
           onSuccess: (response) {
-            print(response.user);
-            print(response.session);
+            context.showSnackBar(
+                message: 'Account registered! Check email for confirmation link.'
+            );
+            onSignUpSuccess();
           },
           metadataFields: [
             MetaDataField(
